python/requirements.txt
requests~=2.28.2
python-dotenv~=0.21.0
supabase~=0.7.1
psycopg2
gitdb~=4.0.10
h11~=0.14.0
dataclasses~=0.6
pip~=22.3.1
wheel~=0.38.4
cryptography~=39.0.0
docutils~=0.19
keyring~=23.13.1
colorama~=0.4.6
anyio~=3.6.2
sniffio~=1.3.0
smmap~=5.0.0
click~=8.1.3
httpcore~=0.16.3
Pygments~=2.14.0
idna~=3.4
rfc3986~=1.5.0
certifi~=2022.12.7
python-gitlab~=3.12.0
twine~=3.8.0
packaging~=23.0
pkginfo~=1.9.6
tqdm~=4.64.1
urllib3~=1.26.14
setuptools~=65.5.0
six~=1.16.0
webencodings~=0.5.1
bleach~=5.0.1
pydantic~=1.10.4
httpx~=0.23.3
gotrue~=0.5.4
invoke~=1.7.3
tomlkit~=0.11.6
python-dateutil~=2.8.2
realtime~=0.0.5
websockets~=10.4
storage3~=0.3.6
deprecation~=2.1.0
supafunc~=0.2.2
zipp~=3.11.0
cffi~=1.15.1
semver~=2.13.0
python-multipart~=0.0.5
SQLAlchemy~=1.4.46
attrs~=22.2.0
aiohttp~=3.8.3